+++ test/PerformanceTests/ManyEntities.cs
@@ -27,8 +27,10 @@ namespace PerformanceTests
 
         public int Get() => this.CurrentValue;
 
+#pragma warning disable DF0307 // DispatchAsync must be used with the entity name, equal to the name of the function it's used in.
         [FunctionName(EntityName)]
         public static Task Run([EntityTrigger] IDurableEntityContext ctx) => ctx.DispatchAsync<ManyEntities>();
+#pragma warning restore DF0307 // DispatchAsync must be used with the entity name, equal to the name of the function it's used in.
 
         [FunctionName(nameof(StartManyEntities))]
         public static async Task<IActionResult> StartManyEntities(